11: Succesfully navigate to the desired page
Some refactoring was done to make the code easier to read. As it was awhile since I last worked on this I found it confusing to go through the process. So I
refactored the code to make it easier to understand. Also, I updated the driver as the old method I was using is becoming obselete.

The next stage requires reading the data off of the page now.