11: Find the daily ratings of the surf location for the week
It can now find the data relating to daily rating of the surf location. This data is returned in the form of a dictionary. This will be further expanded to collect
the surf details, such as the wind speed and swell height for each day. Additionally, some pylint errors were fixed.

Some refactoring was done as in the future of this project I will want it to search for different locations, so a method was created with parameters that can be
changed for flexibility.