22: Using the database information to help navigate the webpages
Using the function created in the database_talker.py file. The main scraping script can now search the website for more surf locations based on the location and
region given. There was a typo in the database that needed to be fixed as it was affecting the find functions used.